requirements.txt
Django
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
jsonfield